refactor: update logger constants to integrate zerolog directly
Refactor the logger module to utilize zerolog's constants for log levels. This change improves consistency by aligning with zerolog's standard logging practices. It simplifies the logger's configuration options and enhances maintainability by removing redundant custom type definitions. Adjustments are made across multiple files including options, constants, and test cases to ensure cohesive integration. The logger now directly references zerolog's log levels instead of custom-defined levels, streamlining the logging process.

Breaking changes:
- Removed custom logging level type and related methods.
