+++ logger/zerolog/writers.go
@@ -11,19 +11,6 @@ import (
 
 const FileNameTimeFormat = "2006-01-02_15_04_05.log"
 
-type closerWriter struct {
-	io.Writer
-	closer func() error
-}
-
-func (w closerWriter) Close() error {
-	return w.closer()
-}
-
-func WithCloser(writer io.Writer, closer func() error) io.WriteCloser {
-	return closerWriter{Writer: writer, closer: closer}
-}
-
 func NewConsoleWriter() io.Writer {
 	return zerolog.NewConsoleWriter(func(w *zerolog.ConsoleWriter) {
 		w.TimeFormat = time.RFC3339
